+++ packages/mt-template-h5/babel.config.js
@@ -1,6 +1,5 @@
 module.exports = {
   plugins: [
-    // require.resolve('react-hot-loader/babel'),
     require.resolve('@babel/plugin-proposal-class-properties'),
   ],
 };