+++ packages/mt-template-h5/src/index.tsx
@@ -1,9 +1,5 @@
-// import { hot } from 'react-hot-loader/root';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
-import 'antd/dist/antd.less';
-
-// const HotApp = hot(App);
 
 ReactDOM.render(<App />, document.getElementById('root'));