+++ packages/mt-template-h5/src/pages/Demo/Demo.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import styles from './style.module.less';
+
+export default function Demo(): JSX.Element {
+  return (
+    <div className={styles.title}>Demo</div>
+  );
+}