+++ packages/mt-template-h5/src/pages/Demo1/index.js
@@ -1,0 +1,3 @@
+import Demo from './Demo1';
+
+export default Demo;