+++ packages/mt-template-h5/src/store.ts
@@ -1,0 +1,9 @@
+import { createContext } from 'react';
+
+export type TypeGlobalContext = {
+};
+
+const GlobalContext = createContext<TypeGlobalContext>({
+});
+
+export default GlobalContext;